+++ README.md
@@ -1,3 +1,1 @@
 # edog_utils
-
-Code and utilities for the eDog desktop quadruped robot
